refactor: fix whole project to use new annotation-based MVC framework
