+++ src/main/java/core/mvc/DispatcherServlet.java
@@ -11,7 +11,6 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
-import java.lang.reflect.InvocationTargetException;
 
 // url 과 서블릿 매핑을 urlPatterns 속성을 통해 해주고 있음.
 // 모든 클라이언트 요청을 받는 서블릿
@@ -23,12 +22,10 @@ public class DispatcherServlet extends HttpServlet {
 
     private static final Logger log = LoggerFactory.getLogger(DispatcherServlet.class);
 
-    private RequestMapping requestMapping;
     private AnnotationHandlerMapping annotationHandlerMapping;
 
     @Override
     public void init() {
-        requestMapping = new RequestMapping();
         annotationHandlerMapping = new AnnotationHandlerMapping("core.mvc");
         annotationHandlerMapping.initialize();
     }
@@ -40,20 +37,9 @@ public class DispatcherServlet extends HttpServlet {
 
         ModelAndView modelAndView;
         HandlerExecution handlerExecution = annotationHandlerMapping.getHandler(req);
-        if(handlerExecution == null){
-            legacyControllerRun(req, resp, url);
-            return;
-        }
         modelAndView = handlerExecution.handle(req, resp);
         modelAndView.getView().render(modelAndView.getModel(), req, resp);
     }
 
-    private void legacyControllerRun(HttpServletRequest req, HttpServletResponse resp, String url) throws IOException, ServletException {
-        ModelAndView modelAndView;
-        log.debug("legacy controller");
-        Controller controller = requestMapping.find(url);
-        modelAndView = controller.execute(req, resp);
-        modelAndView.getView().render(modelAndView.getModel(), req, resp);
-        return;
-    }
+
 }
