+++ src/main/java/core/mvc/HomeController.java
@@ -1,31 +1,28 @@
-package next.controller;
+package core.mvc;
 
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
-import core.mvc.Controller;
-import core.mvc.JspView;
-import core.mvc.ModelAndView;
-import core.mvc.View;
+import core.annotation.Controller;
+import core.annotation.RequestMapping;
 import next.dao.QuestionDao;
 import next.model.Question;
-import next.model.Response;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import java.util.List;
 
-public class HomeController implements Controller {
-    private static final long serialVersionUID = 1L;
+@Controller
+public class HomeController {
     private static final Logger log = LoggerFactory.getLogger(HomeController.class);
 
-    @Override
-    public ModelAndView execute(HttpServletRequest request, HttpServletResponse response) throws IOException {
+    @RequestMapping(value = "/")
+    public ModelAndView home(HttpServletRequest request, HttpServletResponse response) throws IOException {
         QuestionDao questionDao = new QuestionDao();
         List<Question> questions = questionDao.findAll();
         log.debug("questions : {}", questions.toString());
 
         return new ModelAndView(new JspView("home.jsp")).addObject("questions", questions);
     }
+
 }
