--- src/main/java/next/controller/ShowQuestionController.java
@@ -1,28 +1,0 @@
-package next.controller;
-
-import core.mvc.Controller;
-import core.mvc.JspView;
-import core.mvc.ModelAndView;
-import core.mvc.View;
-import next.dao.AnswerDao;
-import next.dao.QuestionDao;
-import next.model.Question;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-
-public class ShowQuestionController implements Controller {
-
-    @Override
-    public ModelAndView execute(HttpServletRequest req, HttpServletResponse res) throws IOException {
-        // questions
-        Long questionId = Long.parseLong(req.getParameter("questionId"));
-        QuestionDao questionDao = new QuestionDao();
-        Question question = questionDao.findById(questionId);
-        req.setAttribute("question", question);
-        // answers
-        AnswerDao answerDao = new AnswerDao();
-        return new ModelAndView(new JspView("/qna/show.jsp")).addObject("answers", answerDao.findByQuestionId(questionId));
-    }
-}
